ai-email-automation-flask/requirements.txt
flask
python-dotenv
google-genai


